Move JobQueryFilters interface from Filters.tsx to types.ts
